feat: Swift package for Apple platforms (Phase 5) (#53)
* feat: Swift package for Apple platforms (Phase 5)

swift/: SPM package wrapping baedeker-ffi for macOS 13+ / iOS 16+:

- BaedekerModule (compilingWasm: / aotArtifact: / aotArtifact()
  serializer) and BaedekerInstance (call, callAsync serialized on a
  per-instance queue, registerHostFunction with Swift closures, fuelLimit)
- WasmValue enum with SIMD16<UInt8> v128; WasmValueType for signatures;
  BaedekerError mapping every status with the runtime's message
- Memory: readMemory/writeMemory (bounds-checked copies),
  withUnsafeMemoryBytes (zero-copy borrow, invalidation documented)
- Host closures via Unmanaged boxes + a C trampoline; thrown Swift
  errors trap the guest with the error description as the message
- build-xcframework.sh: baedeker-ffi for macOS (universal), iOS device,
  iOS simulator (universal) -> Vendor/CBaedeker.xcframework (gitignored)
- XCTest suite: compile/call/host-closure, AOT roundtrip, memory APIs,
  async call, error mapping, host failure
- macOS CI job: build-xcframework.sh + swift build + swift test

Also in this PR (FFI/CLI support for the package):
- baedeker_module_write_aot (C-side AOT serialize; artifact size query
  via null buffer)
- baedeker-cli compile <in.wasm> -o <out.bdkaot> — the build-time half
  of the iOS AOT workflow (documented in swift/README.md)
- BaedekerValue.tag is now a plain uint8_t carrying a BaedekerValueTag:
  fixed-underlying-type enums are C23-only, and a C-int enum field read
  3 bytes of undefined padding against Rust's 1-byte write (found by the
  Linux Swift harness). cbindgen post-process keeps both typedef
  branches enum-typed so Swift's Clang importer sees one candidate.
- build.rs post-processing documented; BaedekerValueTag force-included
  in the header for its constants

Validated end-to-end on Linux (Swift 6.1.2): compile -> AOT roundtrip
-> instantiate -> host closure -> call -> memory -> async -> error
mapping. Apple SPM path validated by the new macOS CI job.

* fix: make build-xcframework.sh executable

* fix: stage module.modulemap in the XCFramework headers (SPM import)

+++ crates/baedeker-ffi/src/api2.rs
@@ -89,7 +89,7 @@ pub unsafe extern "C" fn baedeker_instance_register_host_func(
                 .collect::<Result<_, _>>()?;
             let mut c_results = vec![
                 BaedekerValue {
-                    tag: BaedekerValueTag::I32,
+                    tag: BaedekerValueTag::I32 as u8,
                     data: crate::value::BaedekerValueData { i32_: 0 },
                 };
                 n_results
